+++ DTCD-StyleSystem/src/base-components/BaseSwitch/BaseSwitch.js
@@ -1,8 +1,11 @@
 import html from './BaseSwitch.html';
 
 export default class BaseSwitch extends HTMLElement {
+
+  #checkbox;
+
   static get observedAttributes() {
-    return ['color', 'checked', 'disabled'];
+    return ['checked', 'disabled'];
   }
 
   constructor() {
@@ -13,24 +16,18 @@ export default class BaseSwitch extends HTMLElement {
 
     this.attachShadow({ mode: 'open' });
     this.shadowRoot.appendChild(template.content.cloneNode(true));
-    this.checkbox = this.shadowRoot.querySelector('input');
 
-    this.checkChangeHandler = () => {
-      this.dispatchEvent(new Event('input', { bubbles: true }));
-    };
+    this.#checkbox = this.shadowRoot.querySelector('#check');
 
-    this.changeHandler = e => {
-      e.value = this.checkbox.checked;
-      this.value = this.checkbox.checked;
-    };
-
-    this.checkbox.addEventListener('change', this.checkChangeHandler);
-    this.addEventListener('input', this.changeHandler);
+    this.value = false;
   }
 
-  disconnectedCallback() {
-    this.checkbox.removeEventListener('change', this.checkChangeHandler);
-    this.removeEventListener('change', this.changeHandler);
+  get value() {
+    return this.#checkbox.checked;
+  }
+
+  set value(val) {
+    this.#checkbox.checked = Boolean(val);
   }
 
   get checked() {
@@ -38,11 +35,8 @@ export default class BaseSwitch extends HTMLElement {
   }
 
   set checked(value) {
-    if (value) {
-      this.setAttribute('checked', '');
-    } else {
-      this.removeAttribute('checked');
-    }
+    if (value) this.setAttribute('checked', '');
+    else this.removeAttribute('checked');
   }
 
   get disabled() {
@@ -50,24 +44,17 @@ export default class BaseSwitch extends HTMLElement {
   }
 
   set disabled(value) {
-    if (value) {
-      this.setAttribute('disabled', '');
-    } else {
-      this.removeAttribute('disabled');
-    }
+    if (value) this.setAttribute('disabled', '');
+    else this.removeAttribute('disabled');
   }
 
   attributeChangedCallback(attrName, oldValue, newValue) {
-    if (attrName === 'color') {
-      this.shadowRoot.querySelector('.switch').style.setProperty(`--color`, newValue);
-    }
-
     if (attrName === 'checked') {
-      this.checkbox.checked = this.checked;
+      this.#checkbox.checked = this.checked;
     }
 
     if (attrName === 'disabled') {
-      this.checkbox.disabled = this.disabled;
+      this.#checkbox.disabled = this.disabled;
     }
   }
 }